requirements.txt
requests
python-dotenv
pytest
anthropic
json
